# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HealthPass Data Submission API",
    "description": "Swagger definition for HealthPass data submission API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/datasubmission",
      "description": "CDT"
    }
  ],
  "paths": {
...
    "/mapper/": {
      "post": {
        "tags": [
          "mapper"
        ],
        "summary": "Create mapper",
        "description": "Create mapper",
        "operationId": "createMapper",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "description": "Create mapper",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/mapperRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created mapper",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/mapperRegistrationRequest"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "get": {
        "tags": [
          "mapper"
        ],
        "summary": "Get all mappers",
        "description": "Get all mappers",
        "operationId": "getAllMappers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully get mappers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/allMappers"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "allMappers":{
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/mapperRegistrationRequest"
        }
      },
      "mapperRegistrationRequest": {
        "type": "object",
        "example":{
        "mapperId": "mapperId",
        "mapperName": "mapperName",
        "type":"type",
        "mapper": {
            "resourceType": "resourceType",
            "patient": {
              "resourceType": "resourceType",
              "identifier": {
                "system": "{{credentialSubject.subject.identity[0].system}}",
                "type": "{{credentialSubject.subject.identity[0].type}}",
                "value": "{{credentialSubject.subject.identity[0].value}}"
              },
              "name": [
                {
                  "family": "{{credentialSubject.subject.name.family}}",
                  "given": "{{credentialSubject.subject.name.given}}"
                }
              ],
              "gender": "{{credentialSubject.subject.gender}}",
              "birthDate": "{{credentialSubject.subject.birthDate}}",
              "address": [
                {
                  "text": "{{credentialSubject.subject.address}}"
                }
              ],
              "telecom": [
                {
                  "system": "phone",
                  "value": "{{credentialSubject.subject.phone}}"
                },
                {
                  "system": "email",
                  "value": "{{credentialSubject.subject.email}}"
                }
              ]
            },
            "lotNumber": "{{credentialSubject.lotNumber}}",
            "manufacturer": "{{credentialSubject.manufacturer}}",
            "occurrenceDateTime": "{{credentialSubject.occurrenceDateTime}}",
            "status": "{{credentialSubject.status}}",
            "vaccineCode": "{{credentialSubject.vaccineCode}}"
          }
        }
      },
...
    },
    "responses": {
      "InvalidRequestError": {
        "description": "Missing or invalid required fields in request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
    }
  }
}